Split main.rs: extract activation, lifecycle, power-bar autodetect
Closes #34. Wave 3 architectural reshape from the action plan.

`main.rs` was 605 lines vs CLAUDE.md's "~185 line Coordinator" promise.
Three coherent concerns mashed together; this PR separates them.

Layout after the split:

- `src/main.rs` (92 lines) — true coordinator: arg parsing → log init →
  power-bar autodetect → lifecycle gates → singleton lock → resolve
  paths → build DrawerInit → app.connect_activate → run.
- `src/lifecycle.rs` (78 lines) — `handle_open_close`,
  `handle_existing_instance`, `acquire_singleton_lock`. All run before
  GTK init and may exit the process before any UI work.
- `src/power_bar_detect.rs` (88 lines) — `auto_detect_power_bar` and
  the four private detector helpers (`detect_lock`, `detect_command`,
  `command_on_path`, `can_suspend`). Driven by `--pb-auto` only.
- `src/activate.rs` (423 lines) — GTK activate-time wiring. Owns the
  embedded default CSS (`include_str!`), `DrawerInit` parameter bundle,
  `activate_drawer` itself, and the activate-time helpers
  (`apply_theme_settings`, `apply_force_theme`, `load_preferred_apps`,
  `drawer_monitor_connector`, `resolve_target_monitor`,
  `setup_close_button`). The inline 30-line click-catcher backdrop
  block is now `setup_click_catcher_backdrops` per the issue body.

Also collapses the 9-arg `activate_drawer(app, css_path, config,
app_dirs, compositor, pinned_file, exclusions, data_home, sig_rx)` into
`activate_drawer(app, &DrawerInit)` — the
`#[allow(clippy::too_many_arguments)]` is gone. (The remaining
too_many_arguments allow on `build_app_flow_box` is Issue #36's target.)

Acceptance from #34:
- main.rs ≤ 200 lines ✓ (92)
- CLAUDE.md "What lives where" updated ✓
- #[allow(clippy::too_many_arguments)] on activate_drawer gone ✓

Zero behavior change — pure structural reshape. `make lint` clean.
